requirements.txt
boto3
requests
configparser
bs4
kubernetes
hvac
hvac[parser]